requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
adal==1.2.1               # via botframework-connector
aiohttp==3.6.2
astroid==2.3.3            # via pylint
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp, pytest
botbuilder-core==4.7.0
botbuilder-schema==4.7.0  # via botbuilder-core, botframework-connector
botframework-connector==4.7.0  # via botbuilder-core
certifi==2019.11.28       # via msrest, requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via aiohttp, requests
click==7.0                # via flask
coverage==5.0.1           # via pytest-cov
cryptography==2.8         # via adal, botframework-connector
flask==1.1.1
gunicorn==20.0.4
idna==2.8                 # via requests, yarl
importlib-metadata==1.3.0  # via pluggy, pytest
isodate==0.6.0            # via msrest
isort==4.3.21             # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
jsonpickle==1.2           # via botbuilder-core
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
more-itertools==8.0.2     # via pytest, zipp
msrest==0.6.10            # via botbuilder-schema, botframework-connector
multidict==4.7.3          # via aiohttp, yarl
oauthlib==3.1.0           # via requests-oauthlib
packaging==19.2           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycparser==2.19           # via cffi
pyjwt==1.5.3              # via adal, botframework-connector
pylint==2.4.4
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1
pytest==5.3.2
python-dateutil==2.8.1    # via adal
requests-oauthlib==1.3.0  # via msrest
requests==2.22.0          # via adal, botframework-connector, msrest, requests-oauthlib
six==1.13.0               # via astroid, cryptography, isodate, packaging, python-dateutil
typed-ast==1.4.0          # via astroid
urllib3==1.25.7           # via requests
wcwidth==0.1.8            # via pytest
werkzeug==0.16.0          # via flask
wrapt==1.11.2             # via astroid
yarl==1.4.2               # via aiohttp
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
